--- components/bl-group-member-goal-progress-component/styles/index.less
@@ -1,165 +1,0 @@
-// you can use LESS variables from the current Theme and app extensions of the UI-Container
-@bl-customComponent-groupMemberGoalProgress-themeColor: @themePrimary;
-@bl-customComponent-groupMemberGoalProgress-backgroundColor: @appBackgroundColor;
-@bl-customComponent-groupMemberGoalProgress-textColor: @appTextColor;
-
-@bl-customComponent-groupMemberGoalProgress-chartBarColor: #212121;
-@bl-customComponent-groupMemberGoalProgress-chartNameColor: @bl-customComponent-groupMemberGoalProgress-textColor;
-@bl-customComponent-groupMemberGoalProgress-chartFillColor: @bl-customComponent-groupMemberGoalProgress-themeColor;
-
-@bl-customComponent-groupMemberGoalProgress-labelWidth: 60px;
-@bl-customComponent-groupMemberGoalProgress-labelMarginRight: 16px;
-@bl-customComponent-groupMemberGoalProgress-labelWidthWithOffset: @bl-customComponent-groupMemberGoalProgress-labelWidth + @bl-customComponent-groupMemberGoalProgress-labelMarginRight;
-
-.bl-customComponent-groupMemberGoalProgress {
-  .chart {
-    &__items {
-      position: relative;
-      display: flex;
-      flex-direction: column;
-      justify-content: space-between;
-      height: 100%;
-      width: 100%;
-      margin-bottom: 5.5px;
-    }
-    
-    &__item {
-      position: relative;
-      display: flex;
-      align-items: center;
-      margin-bottom: 16px;
-    }
-    
-    &__name-wrap {
-      width: @bl-customComponent-groupMemberGoalProgress-labelWidth;
-      margin-right: @bl-customComponent-groupMemberGoalProgress-labelMarginRight;
-      flex-shrink: 0;
-    }
-      
-    &__name {
-      color: @bl-customComponent-groupMemberGoalProgress-chartNameColor;
-      font-size: 15px;
-      line-height: 18px;
-    }
-    
-    &__bar {
-      position: relative;
-      background-color: @bl-customComponent-groupMemberGoalProgress-chartBarColor;
-      width: calc(100% - @bl-customComponent-groupMemberGoalProgress-labelWidthWithOffset);
-      height: 20px;
-      cursor: pointer;
-      flex-shrink: 0;
-      
-      &:hover {
-        .chart__shape-container {
-          display: block;
-        }
-      }
-    }
-    
-    &__fill {
-      position: absolute;
-      left: 0;
-      height: 100%;
-      width: 50%;
-      background-color: @bl-customComponent-groupMemberGoalProgress-chartFillColor;
-    }
-    
-    &__shape {
-      &-container {
-        display: none;
-        position: relative;
-        top: 50%;
-        width: 73px;
-        margin-right: auto;
-        margin-left: auto;
-        z-index: 2;
-      }
-      
-      &-triangle {
-        position: absolute;
-        left: 50%;
-        transform: translateX(-50%);
-        width: 0; 
-        height: 0; 
-        border-left: 22px solid transparent;
-        border-right: 22px solid transparent;
-        border-bottom: 22px solid #121212;
-      }
-      
-      &-rectangle {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        position: absolute;
-        top: 16px;
-        width: 73px;
-        height: 47px;
-        color: #FFFFFF;
-        background-color: #121212;
-        border-radius: 10px;
-      }
-    }
-    
-    &__grids {
-      position: absolute;
-      top: 0;
-      right: 0;
-      height: 100%;
-      width: calc(100% - @bl-customComponent-groupMemberGoalProgress-labelWidthWithOffset);
-      z-index: 0;
-    }
-    
-    &__y-grid-line {
-      position: absolute;
-      top: 0;
-      width: 1px;
-      height: 100%;
-      border-left: 1px dashed #525252;
-      transform: translateX(-50%);
-    }
-    
-    &__percentage-marks {
-      position: relative;
-      display: flex;
-      justify-content: space-between;
-      width: calc(100% - @bl-customComponent-groupMemberGoalProgress-labelWidthWithOffset);
-      margin-left: auto;
-      
-      &-item {
-        position: absolute;
-        color: @bl-customComponent-groupMemberGoalProgress-chartNameColor;
-        font-size: 10px;
-        line-height: 11px;
-        transform: translateX(-50%);
-      }
-    }
-  }
-}
-
-@media(min-width: 1201px) {
-  @bl-customComponent-groupMemberGoalProgress-labelWidth: 105px;
-  @bl-customComponent-groupMemberGoalProgress-labelMarginRight: 26px;
-  @bl-customComponent-groupMemberGoalProgress-labelWidthWithOffset: @bl-customComponent-groupMemberGoalProgress-labelWidth + @bl-customComponent-groupMemberGoalProgress-labelMarginRight;
-  
-  .bl-customComponent-groupMemberGoalProgress {
-    .chart {
-      &__name-wrap {
-        width: @bl-customComponent-groupMemberGoalProgress-labelWidth;
-        margin-right: @bl-customComponent-groupMemberGoalProgress-labelMarginRight;
-      }
-      
-      &__bar {
-        width: calc(100% - @bl-customComponent-groupMemberGoalProgress-labelWidthWithOffset);
-      }
-      
-      &__grids {
-        width: calc(100% - @bl-customComponent-groupMemberGoalProgress-labelWidthWithOffset);
-      }
-      
-      &__percentage-marks {
-        width: calc(100% - @bl-customComponent-groupMemberGoalProgress-labelWidthWithOffset);
-      }
-    }
-  }
-}
