+++ components/bl-multiple-bar-chart-component/src/decorate-number.js
