+++ components/bl-multiple-bar-chart-component/src/validate.js
