+++ components/bl-multiple-bar-chart-component/styles/index.less
@@ -1,0 +1,164 @@
+// you can use LESS variables from the current Theme and app extensions of the UI-Container
+@bl-customComponent-multipleBarChart-themeColor: @themePrimary;
+@bl-customComponent-multipleBarChart-backgroundColor: @appBackgroundColor;
+@bl-customComponent-multipleBarChart-textColor: @appTextColor;
+
+@bl-customComponent-multipleBarChart-chartBarColor: #212121;
+@bl-customComponent-multipleBarChart-chartNameColor: @bl-customComponent-multipleBarChart-textColor;
+@bl-customComponent-multipleBarChart-chartFillColor: @bl-customComponent-multipleBarChart-themeColor;
+
+@bl-customComponent-multipleBarChart-labelWidth: 60px;
+@bl-customComponent-multipleBarChart-labelMarginRight: 16px;
+@bl-customComponent-multipleBarChart-labelWidthWithOffset: @bl-customComponent-multipleBarChart-labelWidth + @bl-customComponent-multipleBarChart-labelMarginRight;
+
+.bl-customComponent-multipleBarChart {
+  .chart {
+    &__items {
+      position: relative;
+      display: flex;
+      flex-direction: column;
+      justify-content: space-between;
+      height: 100%;
+      width: 100%;
+      margin-bottom: 5.5px;
+    }
+    
+    &__item {
+      position: relative;
+      display: flex;
+      align-items: center;
+      margin-bottom: 16px;
+    }
+    
+    &__name-wrap {
+      width: @bl-customComponent-multipleBarChart-labelWidth;
+      margin-right: @bl-customComponent-multipleBarChart-labelMarginRight;
+      flex-shrink: 0;
+    }
+      
+    &__name {
+      color: @bl-customComponent-multipleBarChart-chartNameColor;
+      font-size: 15px;
+      line-height: 18px;
+    }
+    
+    &__bar {
+      position: relative;
+      background-color: @bl-customComponent-multipleBarChart-chartBarColor;
+      width: calc(100% - @bl-customComponent-multipleBarChart-labelWidthWithOffset);
+      height: 20px;
+      cursor: pointer;
+      flex-shrink: 0;
+      
+      &:hover {
+        .chart__bar-info {
+          display: block;
+        }
+      }
+      
+      &-fill {
+        position: absolute;
+        left: 0;
+        height: 100%;
+        width: 50%;
+        background-color: @bl-customComponent-multipleBarChart-chartFillColor;
+      }
+    }
+    
+    &__bar-info {
+      display: none;
+      position: relative;
+      top: 50%;
+      width: 100px;
+      margin-left: auto;
+      margin-right: auto;
+      z-index: 2;
+          
+      &-triangle {
+        position: absolute;
+        left: 50%;
+        transform: translateX(-50%);
+        width: 0; 
+        height: 0; 
+        border-left: 22px solid transparent;
+        border-right: 22px solid transparent;
+        border-bottom: 22px solid #121212;
+      }
+          
+      &-rectangle {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        position: absolute;
+        top: 16px;
+        width: 100%;
+        height: 47px;
+        color: #FFFFFF;
+        background-color: #121212;
+        border-radius: 10px;
+      }
+    }
+    
+    &__grids {
+      position: absolute;
+      top: 0;
+      right: 0;
+      height: 100%;
+      width: calc(100% - @bl-customComponent-multipleBarChart-labelWidthWithOffset);
+      z-index: 0;
+    }
+    
+    &__y-grid-line {
+      position: absolute;
+      top: 0;
+      width: 1px;
+      height: 100%;
+      border-left: 1px dashed #525252;
+      transform: translateX(-50%);
+    }
+    
+    &__percentage-marks {
+      position: relative;
+      display: flex;
+      justify-content: space-between;
+      width: calc(100% - @bl-customComponent-multipleBarChart-labelWidthWithOffset);
+      margin-left: auto;
+      
+      &-item {
+        position: absolute;
+        color: @bl-customComponent-multipleBarChart-chartNameColor;
+        font-size: 10px;
+        line-height: 11px;
+        transform: translateX(-50%);
+      }
+    }
+  }
+}
+
+
+@media(min-width: 1201px) {
+  @bl-customComponent-multipleBarChart-labelWidth: 105px;
+  @bl-customComponent-multipleBarChart-labelMarginRight: 26px;
+  @bl-customComponent-multipleBarChart-labelWidthWithOffset: @bl-customComponent-multipleBarChart-labelWidth + @bl-customComponent-multipleBarChart-labelMarginRight;
+  
+  .bl-customComponent-multipleBarChart {
+    .chart {
+      &__name-wrap {
+        width: @bl-customComponent-multipleBarChart-labelWidth;
+        margin-right: @bl-customComponent-multipleBarChart-labelMarginRight;
+      }
+      
+      &__bar {
+        width: calc(100% - @bl-customComponent-multipleBarChart-labelWidthWithOffset);
+      }
+      
+      &__grids {
+        width: calc(100% - @bl-customComponent-multipleBarChart-labelWidthWithOffset);
+      }
+      
+      &__percentage-marks {
+        width: calc(100% - @bl-customComponent-multipleBarChart-labelWidthWithOffset);
+      }
+    }
+  }
+}
